+++ apps/block_scout_web/lib/block_scout_web/plug/logger.ex
@@ -19,48 +19,27 @@ defmodule BlockScoutWeb.Plug.Logger do
   @impl true
   def call(conn, opts) do
     level = Keyword.get(opts, :log, :info)
-    application = Keyword.get(opts, :application, :block_scout_web)
-
-    log(application, conn, level, opts)
 
     start = System.monotonic_time()
 
     Conn.register_before_send(conn, fn conn ->
+      stop = System.monotonic_time()
+      diff = System.convert_time_unit(stop - start, :native, :microsecond)
+      status = Integer.to_string(conn.status)
+
       Logger.log(
         level,
         fn ->
-          stop = System.monotonic_time()
-          diff = System.convert_time_unit(stop - start, :native, :microsecond)
-          status = Integer.to_string(conn.status)
-
-          [connection_type(conn), ?\s, status, " in ", formatted_diff(diff)]
+          [connection_type(conn), ?\s, status, " in ", formatted_diff(diff), " on ", conn.method, ?\s, endpoint(conn)]
         end,
-        opts
+        Keyword.merge(
+          [duration: diff, status: status, unit: "microsecond", endpoint: endpoint(conn), method: conn.method],
+          opts
+        )
       )
 
       conn
     end)
-  end
-
-  defp log(:api, conn, level, opts) do
-    endpoint =
-      if conn.query_string do
-        "#{conn.request_path}?#{conn.query_string}"
-      else
-        conn.request_path
-      end
-
-    Logger.log(level, endpoint, opts)
-  end
-
-  defp log(_application, conn, level, opts) do
-    Logger.log(
-      level,
-      fn ->
-        [conn.method, ?\s, conn.request_path]
-      end,
-      opts
-    )
   end
 
   defp formatted_diff(diff) when diff > 1000, do: [diff |> div(1000) |> Integer.to_string(), "ms"]
@@ -68,4 +47,12 @@ defmodule BlockScoutWeb.Plug.Logger do
 
   defp connection_type(%{state: :set_chunked}), do: "Chunked"
   defp connection_type(_), do: "Sent"
+
+  defp endpoint(conn) do
+    if conn.query_string do
+      "#{conn.request_path}?#{conn.query_string}"
+    else
+      conn.request_path
+    end
+  end
 end